third_party/blink/web_tests/storage/indexeddb/request-continue-abort-expected.txt

Regression test for IDBRequest issue calling continue on a cursor then aborting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "request-continue-abort.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db.createObjectStore('store')


testCursor:
transaction = db.transaction('store', 'readwrite', {durability: 'relaxed'})
store = transaction.objectStore('store')
store.add('value1', 'key1')
store.add('value2', 'key2')


state = 0
request = store.openCursor()


'success' event fired at request.
PASS ++state is 1
request.result.continue()


'error' event fired at request.
PASS ++state is 2


'abort' event fired at transaction.
PASS ++state is 3
PASS successfullyParsed is true

TEST COMPLETE
